refact: separate cmd and pkg
