+++ .gitignore
@@ -1,1 +1,1 @@
-raink
+/raink
